+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Fix `index` to not generate empty module
+* Add dots to `section.chapter-review` in `data-science`
 
 ## [v2.23.0] - 2024-09-09