+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Fix `index` to not generate empty module
+* Add dots to letter answers in `data-science`
 * Add `BakeEmbeddedExerciseQuestion` for baking exercise-like things
 * Bake ordered lists as embedded exercise questions in `computer-science`